+++ sourcecode/CommAudio_Client/CommAudio_Client.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.30723.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CommAudio_Client", "CommAudio_Client\CommAudio_Client.vcxproj", "{67622976-A92C-4142-B241-55D462C5883B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Win32 = Debug|Win32
+		Release|Win32 = Release|Win32
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{67622976-A92C-4142-B241-55D462C5883B}.Debug|Win32.ActiveCfg = Debug|Win32
+		{67622976-A92C-4142-B241-55D462C5883B}.Debug|Win32.Build.0 = Debug|Win32
+		{67622976-A92C-4142-B241-55D462C5883B}.Release|Win32.ActiveCfg = Release|Win32
+		{67622976-A92C-4142-B241-55D462C5883B}.Release|Win32.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
