+++ sourcecode/CommAudio_Client/CommAudio_Client/Debug/CommAudio_Client.tlog/CommAudio_Client.lastbuildstate
@@ -1,0 +1,2 @@
+#TargetFrameworkVersion=v4.0:PlatformToolSet=v120:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit
+Debug|Win32|C:\Users\MrChimick\Documents\Term 4\COMP 4985 - Aman\Comm-Audio\sourcecode\CommAudio_Client\|
