--- sourcecode/CommAudio_Server/CommAudio_Server/driver.cpp
@@ -1,4 +1,0 @@
-int main()
-{
-	return 0;
-}
